cmabio de int a longs

+++ shop-Alejandro-Samuel/src/main/java/es/iesclaradelrey/da2d1e2425/shopalejandrosamuel/controllers/ShopController.java
@@ -36,7 +36,7 @@ public class ShopController {
     }
 
     @GetMapping("/category/{id}")
-    public ModelAndView getAllCategories(@PathVariable int id) {
+    public ModelAndView getAllCategories(@PathVariable Long id) {
 
         Collection<Region> regiones = regionService.findAll();
         Collection<Pokemon> pokemones = pokemonService.findByRegion(id);
