+++ shop-Alejandro-Samuel/src/main/java/es/iesclaradelrey/da2d1e2425/shopalejandrosamuel/services/PokemonService.java
@@ -10,6 +10,6 @@ public interface PokemonService {
     void save(Pokemon t);
     Collection<Pokemon> findAll();
     Optional<Pokemon> findById(Long id);
-    Collection<Pokemon> findByRegion(int id);
-    Collection<Pokemon> findByType(int id);
+    Collection<Pokemon> findByRegion(Long id);
+    Collection<Pokemon> findByType(Long id);
 }
