+++ shop-Alejandro-Samuel/src/main/java/es/iesclaradelrey/da2d1e2425/shopalejandrosamuel/services/RatingService.java
@@ -9,6 +9,6 @@ public interface RatingService {
     long count();
     void save(Rating t);
     Collection<Rating> findAll();
-    Optional<Rating> findById(int id);
+    Optional<Rating> findById(Long id);
     Collection<Rating> findByPokemon_Id(Long pokemonId);
 }
